fix props interfaces and test in TS app
