+++ jambonz-ui/test-app/src/app.tsx
@@ -1,0 +1,31 @@
+import React, { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import { BrowserRouter, Routes, Route, Outlet } from 'react-router-dom';
+
+import HomePage from './home';
+
+import '../../pkg/public/css/styles.css';
+import './app.css';
+
+const AppLayout = () => (
+  <Outlet />
+);
+
+const App = () => {
+  return (
+    <StrictMode>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<AppLayout />}>
+            <Route index element={<HomePage />} />
+          </Route>
+        </Routes>
+      </BrowserRouter>
+    </StrictMode>
+  );
+};
+
+export const elem = document.getElementById('root') as HTMLElement;
+export const root = createRoot(elem);
+
+root.render(<App />);
